relocated function to mixin for reuse
